Create CodeThatTellsYouAgeOfPlant
Hi, this is just some code I which tells you the age of plant via VS Code. I also think this code could be used as the basis for accuracy testing: So you know how multiple boxes come up when you put in an image such as harvest, 2 weeks old etc. I think using this code we could essentially calculate the area of each box and display the largest box as the answer. Right now, what is happening in that it uses a dictionary function. This means it separates parts of the image data into for example, class (the word) and harvest (the definition), and does this for every single box it detects on the screen. So I'm thinking we make a function to calculate the area of all the boxes and displays the largest one as the answer to the viewer (right now the answer being displayed is just the first box it detects). Anyway, I hope you like what I did so far.. it took a long time to separate all the data and display like this.